+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## dev-master
 
-* #1: `append($item)` methods return `$item` for chains
+* #2: `append($item)` methods return `$item` for chains
+* #3: the directive name can be NULL (it leads to disable directive)
 
 ## 0.1.0 (11 June 2021)